refactor(hub): move data-subject rights → with-audit/portability; tab-* → with-party
git mv the GDPR export/withdraw ops (export-accessible, withdraw-accessible,
request-withdrawal) into a new src/with-audit/portability/ folder, and the
tab-coordination / tab-write-relay files into src/with-party/. Pure file
moves plus import-path rewrites; no behavior change.
