+++ packages/hub/__tests__/export-accessible.test.ts
@@ -7,7 +7,7 @@ import { describe, it, expect } from 'vitest'
 import type { NoydbStore, EncryptedEnvelope, VaultSnapshot } from '../src/types.js'
 import { ConflictError } from '../src/errors.js'
 import { createNoydb } from '../src/noydb.js'
-import { readNoydbBundle } from '../src/with-share/bundle/bundle.js'
+import { readNoydbBundle } from '../src/with-pod/bundle.js'
 
 function makeStore(): NoydbStore {
   const store = new Map<string, Map<string, Map<string, EncryptedEnvelope>>>()
