+++ packages/hub/__tests__/tab-coordination.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it } from 'vitest'
-import { TabCoordinator, type TabLockManager, type TabChannel } from '../src/tab-coordination.js'
+import { TabCoordinator, type TabLockManager, type TabChannel } from '../src/with-party/tab-coordination.js'
 
 /** FIFO exclusive lock manager (mirrors by-peer's createMockLocks). */
 function mockLocks(): TabLockManager {
