+++ packages/hub/__tests__/tab-write-propagation.test.ts
@@ -2,7 +2,7 @@ import { describe, expect, it } from 'vitest'
 import { createNoydb } from '../src/noydb.js'
 import { withHistory } from '../src/with-commit/history/index.js'
 import { memory } from '../../to-memory/src/index.js'
-import type { TabChannel } from '../src/tab-coordination.js'
+import type { TabChannel } from '../src/with-party/tab-coordination.js'
 
 /** In-memory broadcast bus (each send reaches all OTHER channels). */
 function makeBus(n: number): TabChannel[] {
